feat: Remove unused dependencies and refactor request parameters
This commit removes unused dependencies from the 'pnpm-lock.yaml' file and the 'package.json' file. It also refactors the request parameters in the 'api.ts' file according to the Swagger documentation. This includes changing the 'body' parameter to 'header' and 'x-session-id' and 'x-access-token' parameters for the API routes. It also updates the Swagger documentation file to reflect these changes. The version in 'package.json' is also bumped up.